% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_sgcca.R
\name{analyze}
\alias{analyze}
\title{Analyze a sgcca object}
\usage{
analyze(sgcca)
}
\arguments{
\item{sgcca}{SGCCA object from the RGCCA package}
}
\value{
A vector with the correlation between components, AVE (both inner
and outer), the canonical correlation, the weight in the design matrix, and
the number of interactions that exists.
}
\description{
Calculates the correlation between the first components to reach the
canonical correlation as well.
}
\examples{
data("Russett", package = "RGCCA")
X_agric <- as.matrix(Russett[, c("gini", "farm", "rent")])
X_ind <- as.matrix(Russett[, c("gnpr", "labo")])
X_polit <- as.matrix(Russett[ , c("inst", "ecks",  "death", "demostab",
                                  "dictator")])
A <- list(X_agric, X_ind, X_polit)
A <- lapply(A, function(x) RGCCA::scale2(x, bias = TRUE))
C <- matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3)
out <- RGCCA::rgcca(A, C, tau =rep(0, 3), scheme = "factorial",
                    scale = FALSE, verbose = TRUE)
analyze(out)
}
