% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_sgcca.R
\name{aves}
\alias{aves}
\title{Method to simplify AVE}
\usage{
aves(x)
}
\arguments{
\item{x}{rgcca or sgcca object}
}
\value{
The same object with AVE_X simplified
}
\description{
This simplifies the AVE_X to make it easier to understand
}
