% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_design.R
\name{boot_design}
\alias{boot_design}
\title{Explore multiple designs}
\usage{
boot_design(
  designs,
  sample = 10000,
  ...,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{designs}{A list of designs such as that obtained with [weight_design()].}

\item{sample}{A numeric value with the number of samples to test.}

\item{...}{
  Arguments passed on to \code{\link[=boot_samples_sgcca]{boot_samples_sgcca}}
  \describe{
    \item{\code{nb_boot}}{Number of bootstraps to perform.}
    \item{\code{verbose}}{Logical, should it print a progress bar (default) or not?}
  }}

\item{BPPARAM}{Set up parallel backend (see BiocParallel documentation).}
}
\value{
A data.frame with information about the models performance.
}
\description{
Explore multiple designs
}
\seealso{
[boot_index()]
}
