% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_rgcca.R
\name{boot_index}
\alias{boot_index}
\title{Create index for the samples}
\usage{
boot_index(samples, boots)
}
\arguments{
\item{samples}{A numeric value of samples.}

\item{boots}{A numeric value of bootstraps you want.}
}
\value{
A list of length [boots] with indices for the samples.
}
\description{
Helper function for the easy creation of an index to bootstrap.
Each element has the same number of samples.
}
\examples{
boot_index(10, 3)
}
\seealso{
[boot_index_sgcca()]
}
