% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.R
\name{cca_rgcca}
\alias{cca_rgcca}
\title{Canonical correlations and covariates}
\usage{
cca_rgcca(rgcca)
}
\arguments{
\item{rgcca}{The output of SGCCA or RGCCA}
}
\value{
A list of matrix with the correlation and covariation between CCA
dimensions
}
\description{
Calculates the canonical correlation and canonical covariates.
}
