% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correct}
\alias{correct}
\title{Check that the design is fully connected}
\usage{
correct(x)
}
\arguments{
\item{x}{Design matrix, a symmetric matrix with weights between 0 and 1.}
}
\value{
A logical value if it is fully connected or not.
}
\description{
Given the design matrix, checks that all the blocks are connected between them.
Without leaving any block not connected or some blocks isolated from the others.
}
\examples{
C <- matrix(0, nrow = 4, ncol = 4, dimnames = list(LETTERS[1:4], LETTERS[1:4]))
(m1 <- subSymm(C, 1, 2, 1))
correct(m1) # Several blocks wouldn't be connected
(m2 <- subSymm(m1, 3, 4, 1))
correct(m2) # All blocks are connected, but there are two networks.
(m3 <- subSymm(m2, 1, 4, 1))
correct(m3) # All blocks are connected and there is a single network
}
\references{
See this question and answer about how it is done:
\url{https://math.stackexchange.com/a/551947}
}
