% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSymm.R
\name{extract_model}
\alias{extract_model}
\title{Extract model}
\usage{
extract_model(model, models_eval, which)
}
\arguments{
\item{model}{A matrix with the names of the blocks.}

\item{models_eval}{The output of \code{iterate_model()} or \code{search_model()}.}

\item{which}{Either inner, outer or the row to extract the model.}
}
\value{
A matrix with the weights on the models_eval argument.
}
\description{
Once evaluated extract a model.
}
\examples{
data("Russett", package = "RGCCA")
X_agric <- as.matrix(Russett[, c("gini", "farm", "rent")])
X_ind <- as.matrix(Russett[, c("gnpr", "labo")])
X_polit <- as.matrix(Russett[ , c("inst", "ecks",  "death", "demostab",
                                  "dictator")])
A <- list(Agric = X_agric, Ind = X_ind, Polit = X_polit)
C <- matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3)
out <- search_model(A = A, C = C, c1 =rep(1, 3), scheme = "factorial",
                    scale = FALSE, verbose = FALSE,  ncomp = rep(1, length(A)),
                    bias = TRUE, BPPARAM = BiocParallel::SerialParam())
extract_model(C, out, "inner")
}
