% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_sgcca.R
\name{improve}
\alias{improve}
\title{Improve the information on sgcca classes}
\usage{
improve(sgcca, namesA)
}
\arguments{
\item{sgcca}{An object of class \code{sgcca}.}

\item{namesA}{The names of the original data}
}
\value{
An object of class sgcca
}
\description{
Add names to data, simplify AVE output
}
