% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteRmodel-package.R
\docType{package}
\name{inteRmodel-package}
\alias{inteRmodel}
\alias{inteRmodel-package}
\title{inteRmodel: Find Models of Interaction}
\description{
Provides functions to be used on generalized canonical correlation analysis. For assessing the stability of the model, finding the right model and evaluating the models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/llrs/inteRmodel}
  \item Report bugs at \url{https://github.com/llrs/inteRmodel/issues}
}

}
\author{
\strong{Maintainer}: Lluís Revilla Sancho \email{lluis.revilla@gmail.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID})

}
\keyword{internal}
