% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate}
\alias{iterate}
\title{Iterate to find the best model}
\usage{
iterate(...)
}
\arguments{
\item{...}{All the same arguments that would be passed to sggca, pass named arguments.}
}
\description{
Given a design and
}
\examples{
data("Russett", package = "RGCCA")
X_agric <- as.matrix(Russett[, c("gini", "farm", "rent")])
X_ind <- as.matrix(Russett[, c("gnpr", "labo")])
X_polit <- as.matrix(Russett[ , c("inst", "ecks",  "death", "demostab",
                                  "dictator")])
A <- list(X_agric, X_ind, X_polit)
A <- lapply(A, function(x) RGCCA::scale2(x, bias = TRUE))
C <- matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3)
out <- iterate(A = A, C = C, c1 =rep(1, 3), scheme = "factorial",
               scale = FALSE, verbose = FALSE,
               ncomp = rep(2, length(A)),
               bias = TRUE)
head(out)
}
\seealso{
[sgcca]
}
