% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{looIndex}
\alias{looIndex}
\title{Indexes without one sample}
\usage{
looIndex(size)
}
\arguments{
\item{size}{The number of samples}
}
\value{
A list of indices where one sample is excluded
}
\description{
Performs the leave-out-one indexing
}
\examples{
looIndex(15)
}
