% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale2}
\alias{scale2}
\title{Scale2}
\usage{
scale2(...)
}
\arguments{
\item{...}{Named arguments for scale2, A (Matrix), center, scale, bias (logical).}
}
\description{
If possible reexport scale2 of RGCCA, if not make it available to users.
}
