% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetData.R
\name{subsetData}
\alias{subsetData}
\title{Subset a list}
\usage{
subsetData(A, index)
}
\arguments{
\item{A}{A list of an array with samples in rows and variables in the columns.}

\item{index}{The samples to keep.}
}
\value{
A list with the samples.
}
\description{
Given a list and an index subset each element of the list and
remove the variables that are constant.
}
\note{
It also removes constant variables of the data.
}
\examples{
data(ge_cgh_locIGR)
A <- subsetData(ge_cgh_locIGR$multiblocks, sample(53, repe))
str(A)
}
