% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSymm.R
\name{subSymm}
\alias{subSymm}
\title{Substitute in a symmetric matrix}
\usage{
subSymm(m, x, y, val)
}
\arguments{
\item{m}{The symmetric matrix}

\item{x}{Row position}

\item{y}{Column position}

\item{val}{Value to insert in the given position}
}
\value{
The symmetric matrix with the value inserted in the right positions
}
\description{
Helper to create models, which substitute in a symmetric matrix for a value.
}
\examples{
m <- matrix(0, ncol = 3, nrow = 3)
subSymm(m, 1, 2, 1)
}
\seealso{
\code{\link{symm}}, \code{\link{correct}}
}
