% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSymm.R
\name{symm}
\alias{symm}
\title{Create symmetric matrix}
\usage{
symm(m, data)
}
\arguments{
\item{m}{Square numeric matrix.}

\item{data}{Numeric values of the upper triangular side of the matrix}
}
\value{
A square symmetric matrix.
}
\description{
Create symmetric matrix
}
\note{
After the upper case there can be other values that are assumed to be
in the diagonal.
}
\examples{
m <- matrix(0, ncol = 5, nrow = 5)
symm(m, c(0, 1, 1, 1, 0, 0.5, 0.2, 0.3, 0.7, 0.1))
}
\seealso{
\code{\link{subSymm}}, \code{\link{correct}}
}
