% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_variables.R
\name{variables_relations}
\alias{variables_relations}
\title{Find the variables that are related}
\usage{
variables_relations(rel, comp = c(1, 2))
}
\arguments{
\item{rel}{Relationship model}

\item{comp}{A numeric vector with the blocks you want to compare.}
}
\description{
Performs the correlation between the variables of the block.
}
\examples{
\dontrun{
# Download the dataset's package at http://biodev.cea.fr/sgcca/.
# --> gliomaData_0.4.tar.gz

require(gliomaData)
data(ge_cgh_locIGR)

A <- ge_cgh_locIGR$multiblocks
Loc <- factor(ge_cgh_locIGR$ylabel)
levels(Loc) <- colnames(ge_cgh_locIGR$multiblocks$y)
C <-  matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3)
tau = c(1, 1, 0)

# rgcca algorithm using the dual formulation for X1 and X2
# and the dual formulation for X3
A[[3]] = A[[3]][, -3]
# sgcca algorithm
result.sgcca = sgcca(A, C, c1 = c(.071,.2, 1), ncomp = c(1, 1, 1),
                     scheme = "centroid", verbose = FALSE)
weights <- variables_contribution(A, result.sgcca)
rel <- variables_relations(weights)
rel[1:15, 1:5]
}
}
