% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_design.R
\name{weight_design}
\alias{weight_design}
\title{Create design matrices with different weights}
\usage{
weight_design(weights = 4, size, diff0 = NULL)
}
\arguments{
\item{weights}{A numeric value with the number of weights to be used}

\item{size}{A numeric value with the number of datasets on the design.}

\item{diff0}{A Numeric vector of position which should be different from 0
from the \code{lower.tri()}. See examples}
}
\value{
A list of matrices with the designs with different weights.
}
\description{
Creates all the possible matrices given a number of blocks and a number of
weights. Optionally it can create just a subset of those based on a numeric
vector.
}
\examples{
out <- weight_design(4, 4)
head(out)
# Using previously defined matrix
C <- diag(4)
diag(C) <- 0
C <- subSymm(C, 1, 2, 0.456)
C <- subSymm(C, 2, 3, runif(1))
C <- subSymm(C, 1, 3, runif(1))
C <- subSymm(C, 3, 4, runif(1))
# Explore matrices which are similar to C
d <- weight_design(weights = 11, size = 4, which(lower.tri(C) & C != 0))
head(d)
}
\seealso{
\code{subSymm()}
}
\author{
Flodel: \url{https://codereview.stackexchange.com/a/203517/36067}
}
