% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_sgcca.R
\name{analyze}
\alias{analyze}
\title{Analyze a sgcca object}
\usage{
analyze(sgcca)
}
\arguments{
\item{sgcca}{SGCCA object from the RGCCA package}
}
\value{
A vector with the correlation between components, AVE (both inner
and outer), the canonical correlation, the weight in the design matrix, and
the number of interactions that exists.
}
\description{
Calculates the correlation between the first components to reach the
canonical correlation as well.
}
