% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_sgcca.R
\name{aves}
\alias{aves}
\title{Method to simplify AVE}
\usage{
aves(x)
}
\arguments{
\item{x}{rgcca or sgcca object}
}
\value{
The same object with AVE_X simplified
}
\description{
This simplifies the AVE_X to make it easier to understand.
}
\details{
It converts to a matrix where each row is a dimension and each column is a block.
}
\examples{
data("Russett", package = "RGCCA")
X_agric <- as.matrix(Russett[, c("gini", "farm", "rent")])
X_ind <- as.matrix(Russett[, c("gnpr", "labo")])
X_polit <- as.matrix(Russett[ , c("inst", "ecks",  "death", "demostab",
                                  "dictator")])
A <- list(X_agric, X_ind, X_polit)
A <- lapply(A, function(x) RGCCA::scale2(x, bias = TRUE))
C <- matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3)
out <- RGCCA::rgcca(A, C, tau =rep(0, 3), scheme = "factorial",
                    scale = FALSE, verbose = FALSE, ncomp = rep(2, length(A)))
out$AVE
out <- aves(out)
out$AVE
}
