% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_rgcca.R
\name{boot_sgcca}
\alias{boot_sgcca}
\title{Bootstrap sgcca}
\usage{
boot_sgcca(A, C, shrinkage, nb_boot = 1000)
}
\arguments{
\item{A}{The list with the original data}

\item{C}{The symmetric matrix with the relationships between datsets.}

\item{shrinkage}{Shrinkage estimated (use the estimated for the original datastet)}

\item{nb_boot}{Number of bootstraps to perform}
}
\value{
A list with two elements: the coefficient of each variable of the
input blocks; and the AVE values, both inner, and outer
}
\description{
Performs the centroid bootstrap
}
