% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.R
\name{cca_rgcca}
\alias{cca_rgcca}
\title{Canonical correlations and covariates}
\usage{
cca_rgcca(rgcca)
}
\arguments{
\item{rgcca}{The output of SGCCA or RGCCA}
}
\value{
A data.frame with the correlation and covariation between CCA
components
}
\description{
Calculates the canonical correlation and canonical covariates.
}
\examples{
data("Russett", package = "RGCCA")
X_agric <- as.matrix(Russett[, c("gini", "farm", "rent")])
X_ind <- as.matrix(Russett[, c("gnpr", "labo")])
X_polit <- as.matrix(Russett[ , c("inst", "ecks",  "death", "demostab",
                                  "dictator")])
A <- list(X_agric, X_ind, X_polit)
A <- lapply(A, function(x) scale2(x, bias = TRUE))
C <- matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3)
out <- RGCCA::rgcca(A, C, tau =rep(0, 3), scheme = "factorial",
                    scale = FALSE, verbose = FALSE, ncomp = rep(2, length(A)))
out <- improve(out, c("Agric", "Ind", "Polit"))
ccas <- cca_rgcca(out)
}
