% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correct}
\alias{correct}
\title{Check that the design is fully connected}
\usage{
correct(x)
}
\arguments{
\item{x}{Design matrix, a symmetric matrix with}
}
\value{
A logical value if it is fully connected or not.
}
\description{
Given the design matrix, checks that all the blocks are connected between them.
Without leaving any block not connected or some blocks isolated from the others.
}
\references{
\url{https://math.stackexchange.com/a/551947}
}
