% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_sgcca.R
\name{improve}
\alias{improve}
\title{Improve the information on sgcca classes}
\usage{
improve(sgcca, namesA)
}
\arguments{
\item{sgcca}{An object of class \code{sgcca} or \code{rgcca}.}

\item{namesA}{The names of the original data}
}
\value{
An object of class \code{sgcca} or \code{rgcca}
}
\description{
Add names to data returned by \code{rgcca} or \code{sgcca}, simplify AVE output
}
\examples{
data("Russett", package = "RGCCA")
X_agric <- as.matrix(Russett[, c("gini", "farm", "rent")])
X_ind <- as.matrix(Russett[, c("gnpr", "labo")])
X_polit <- as.matrix(Russett[ , c("inst", "ecks",  "death", "demostab",
                                  "dictator")])
A <- list(X_agric, X_ind, X_polit)
A <- lapply(A, function(x) scale2(x, bias = TRUE))
C <- matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3)
out <- RGCCA::rgcca(A, C, tau = rep(0, 3), scheme = "factorial",
                    scale = FALSE, verbose = FALSE, ncomp = rep(2, length(A)))
out$AVE
out <- improve(out, c("Agric", "Ind", "Polit"))
out$AVE
}
