% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate}
\alias{iterate}
\title{Iterate to find the best model}
\usage{
iterate(...)
}
\arguments{
\item{...}{All the same arguments that would be passed to sggca, pass named arguments.}
}
\description{
Given a design and
}
\seealso{
[sgcca]
}
